Extract get-foreground
