+++ packages/alfa-rules/src/common/dom/get-colors.ts
@@ -4,4 +4,5 @@
 export * from "./get-colors/color";
 export * from "./get-colors/color-error";
 export * from "./get-colors/get-colors";
+export * from "./get-colors/get-foreground";
 export * from "./get-colors/get-layers";
